Move logger to util folder
